composedb/app/requirements.txt
anyio==3.6.2
autopep8==2.0.0
click==8.1.3
fastapi==0.86.0
h11==0.14.0
idna==3.4
pycodestyle==2.9.1
pydantic==1.10.2
python-dotenv==0.21.1
sniffio==1.3.0
starlette==0.20.4
tomli==2.0.1
typing-extensions==4.4.0
uvicorn==0.19.0
